Moved two utils files to a sub-folder
