+++ utils/validation/checkIfEmpty.ts
