+++ utils/validation/removeWhitespace.ts
